requirements.txt
# SHL Assessment Recommender — Requirements
# Install: pip install -r requirements.txt

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1

# Embeddings (runs locally — free)
sentence-transformers==3.0.1

# Vector DB — Pinecone Cloud
pinecone==4.1.0

# LLM
google-generativeai==0.7.2

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0

# Utils
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.4